RAVE 0.504
- Solucionat tema del manifest que nomes em deixava detectar el windows fins a la 8.0, ara ja detecta el windows 10.
- Solucionat el des-ordre de finestres al mostrar un tooltipinfo, aixo pasava quan obria les opcions, i es mostrava un tooltip del player.
- DBarraScrollEx machacada... per començar es basava en un scroll del 0% al 100% (lo que s'habia de restar la pàgina), ara va de 0.0 a 1.0 i la pagina ja no s'ha de restar (es computa abans de arrivar al 1.0)
- Ara la DBarraScrollEx te un minim de 10 pixels per la barra (si hi han 1000000000 items, seguira mostran la barra de 10 pixels)
- Afegides transicions a DBarraScrollEx, i ames s'ha modificat el estat Resaltado per : BarraResaltada i FondoResaltado.
- Modificats DListaEx i DArbolEx per calcular bé la pàgina del DBarraScrollEx
- Retocats els colors de DBarraScrollEx per que resaltin un pel mes.
- DMenuEx ara es comproba que hi hagi suficient espai per mostrar el menú, i si no n'hi ha el mostro adaptat al monitor (per que mai es surti del monitor)

+++ Reproductor de Audio y Video Extendido/ListaMedios.h
@@ -26,8 +26,8 @@ class ListaMedios : public DWL::DListaEx {
 
 	ItemMedio                  *BuscarHash(sqlite3_int64 Hash);
 
-	inline ItemMedio           *Medio(const LONGLONG nPos)	{ return static_cast<ItemMedio *>(_Items[nPos]); 												}
-	inline ItemMedio           *MedioMarcado(void)			{ return static_cast<ItemMedio *>(ItemMarcado());												}
+	inline ItemMedio           *Medio(const LONGLONG nPos)	{ return static_cast<ItemMedio *>(_Items[static_cast<unsigned int>(nPos)]); 											}
+	inline ItemMedio           *MedioMarcado(void)			{ return static_cast<ItemMedio *>(ItemMarcado());																		}
 
 	inline ItemMedio		   *MedioResaltado(void)		{ return static_cast<ItemMedio *>(ItemResaltado());																		}
 	inline ItemMedio           *UMedioResaltado(void)		{ return static_cast<ItemMedio *>((_ItemUResaltado == -1) ? NULL : _Items[static_cast<unsigned int>(_ItemUResaltado)]);	}
